23.01.16 23:33 이현도 - HTTPSession 반영 완료, exception 처리 완료

+++ src/main/java/com/b4/apollo/blog/controller/BlogController.java
@@ -1,5 +1,6 @@
 package com.b4.apollo.blog.controller;
 
+import com.b4.apollo.blog.exception.CommonException;
 import com.b4.apollo.blog.model.dto.BlogDTO;
 import com.b4.apollo.blog.model.dto.BlogForm;
 import com.b4.apollo.blog.model.dto.CommentDTO;
@@ -15,6 +16,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.io.IOException;
 import java.util.List;
@@ -47,12 +49,27 @@ public class BlogController {
     }
 
     @PostMapping("/create")
-    public String questionCreate(@Valid BlogForm blogForm, BindingResult bindingResult, MultipartFile file) throws IOException {
+    public String blogCreate(HttpSession session, @Valid BlogForm blogForm, BindingResult bindingResult, MultipartFile file) throws Exception {
         if (bindingResult.hasErrors()) {
             return "/blog/blogForm";
         }
-        blogService.insertBlog(blogForm.getUserId(), blogForm.getBlogTitle(), blogForm.getBlogContent(), file);
-        return "redirect:/blog/list";
+
+        BlogDTO blog = new BlogDTO();
+
+        String reporter = (String) session.getAttribute("userId");
+
+        try {
+            if (reporter.equals("admin")){
+                blogService.insertBlog(reporter, blogForm.getBlogTitle(), blogForm.getBlogContent(), file);
+                return "redirect:/blog/list";
+            }
+        }catch(NullPointerException e){
+            throw new CommonException("관리자만 작성 가능합니다.");
+        } catch(Exception e){
+            e.printStackTrace();
+            throw new CommonException("에러 발생");
+        }
+        return "redirect:/main";
     }
 
      //질문 게시판 상세 조회
@@ -68,28 +85,57 @@ public class BlogController {
 
     // 질문 수정
     @GetMapping("/modify/{blogNo}")
-    public String questionModify(BlogForm blogForm, @PathVariable("blogNo") int blogNo) {
+    public String blogModify(HttpSession session, BlogForm blogForm, @PathVariable("blogNo") int blogNo) {
         BlogDTO blog = this.blogService.selectBlog(blogNo);
+
+        String reporter = (String) session.getAttribute("userId");
+
+        blog.setReporter(reporter);
         blogForm.setBlogTitle(blog.getBlogTitle());
         blogForm.setBlogContent(blog.getBlogContent());
         return "/blog/blogForm";
     }
 
     @PostMapping("/modify/{blogNo}")
-    public String questionModify(@Valid BlogForm blogForm, BindingResult bindingResult,
+    public String blogModify(HttpSession session, @Valid BlogForm blogForm, BindingResult bindingResult,
                                  @PathVariable("blogNo") int blogNo, MultipartFile file) throws IOException {
         if (bindingResult.hasErrors()) {
             return "/blog/blogForm";
         }
         BlogDTO blog = this.blogService.selectBlog(blogNo);
-        this.blogService.updateBlog(blog, blogForm.getBlogTitle(), blogForm.getBlogContent(), file);
-        return String.format("redirect:/blog/detail/%s", blogNo);
+
+        String reporter = (String) session.getAttribute("userId");
+
+        try{
+            if(reporter.equals("admin")){
+                this.blogService.updateBlog(blog, blogForm.getBlogTitle(), blogForm.getBlogContent(), file);
+                return String.format("redirect:/blog/detail/%s", blogNo);
+            }
+        }catch(NullPointerException e){
+            throw new CommonException("관리자만 수정 가능합니다.");
+        }catch(Exception e) {
+            throw new CommonException("에러 발생");
+        }
+        return "redirect:/main";
     }
 
     @GetMapping("/delete/{blogNo}")
-    private String deleteBlog(@PathVariable("blogNo")  int blogNo) {
-        blogService.deleteBlog(blogNo);
-        return "redirect:/blog/list";
-    }
+    private String deleteBlog(HttpSession session, @PathVariable("blogNo")  int blogNo) {
 
+        BlogDTO blog = this.blogService.selectBlog(blogNo);
+
+        String reporter = (String) session.getAttribute("userId");
+
+        try {
+            if (reporter.equals("admin")) {
+                blogService.deleteBlog(blogNo);
+                return "redirect:/blog/list";
+            }
+        } catch(NullPointerException e){
+            throw new CommonException("관리자만 삭제 가능합니다.");
+        } catch(Exception e) {
+            throw new CommonException("에러 발생");
+        }
+        return "redirect:/main";
+    }
 }
